+++ test/unit_tests/models/test_models.py
@@ -13,7 +13,7 @@ import pytest
 from braindecode.models import (
     Deep4Net, EEGNetv4, EEGNetv1, HybridNet, ShallowFBCSPNet, EEGResNet, TCN,
     SleepStagerChambon2018, SleepStagerBlanco2020, SleepStagerEldele2021, USleep,
-    EEGITNet, EEGInception, TIDNet)
+    EEGITNet, EEGInception, EEGInceptionERP, EEGInceptionMI, TIDNet)
 
 from braindecode.util import set_random_seeds
 
@@ -121,8 +121,9 @@ def test_eegitnet(input_sizes):
     check_forward_pass(model, input_sizes,)
 
 
-def test_eeginception(input_sizes):
-    model = EEGInception(
+@pytest.mark.parametrize("model_cls", [EEGInception, EEGInceptionERP])
+def test_eeginception_erp(input_sizes, model_cls):
+    model = model_cls(
         n_classes=input_sizes['n_classes'],
         in_channels=input_sizes['n_channels'],
         input_window_samples=input_sizes['n_in_times'])
@@ -130,11 +131,12 @@ def test_eeginception(input_sizes):
     check_forward_pass(model, input_sizes,)
 
 
-def test_eeginception_n_params():
+@pytest.mark.parametrize("model_cls", [EEGInception, EEGInceptionERP])
+def test_eeginception_erp_n_params(model_cls):
     """Make sure the number of parameters is the same as in the paper when
     using the same architecture hyperparameters.
     """
-    model = EEGInception(
+    model = model_cls(
         in_channels=8,
         n_classes=2,
         input_window_samples=128,  # input_time
@@ -147,6 +149,47 @@ def test_eeginception_n_params():
 
     n_params = sum(p.numel() for p in model.parameters() if p.requires_grad)
     assert n_params == 14926  # From paper's TABLE IV EEG-Inception Architecture Details
+
+
+def test_eeginception_mi(input_sizes):
+    sfreq = 250
+    model = EEGInceptionMI(
+        n_classes=input_sizes['n_classes'],
+        in_channels=input_sizes['n_channels'],
+        input_window_s=input_sizes['n_in_times'] / sfreq,
+        sfreq=sfreq,
+    )
+
+    check_forward_pass(model, input_sizes,)
+
+
+@pytest.mark.parametrize(
+    "n_filter,reported",
+    [(6, 51386), (12, 204002), (16, 361986), (24, 812930), (64, 5767170)]
+)
+def test_eeginception_mi_binary_n_params(n_filter, reported):
+    """Make sure the number of parameters is the same as in the paper when
+    using the same architecture hyperparameters.
+
+    Note
+    ----
+    For some reason, we match the correct number of parameters for all
+    configurations in the binary classification case, but none for the 4-class
+    case... Should be investigated by contacting the authors.
+    """
+    model = EEGInceptionMI(
+        in_channels=3,
+        n_classes=2,
+        input_window_s=3.,  # input_time
+        sfreq=250,
+        n_convs=3,
+        n_filters=n_filter,
+        kernel_unit_s=0.1,
+    )
+
+    n_params = sum(p.numel() for p in model.parameters() if p.requires_grad)
+    # From first column of TABLE 2 in EEG-Inception paper
+    assert n_params == reported
 
 
 @pytest.mark.parametrize(